requirements.txt
drf-spetacular==0.26.4
djangorestframework==3.14.0